requirements.txt
streamlit
transformers==4.41.2
torch
deepface
tf-keras
tensorflow
matplotlib
pillow
numpy